### 2020-05-23 #2244 @whedon
[PRE REVIEW]: geemap: A Python package for interactive mapping with Google Earth Engine
**Submitting author:** @giswqs (<a href="http://orcid.org/0000-0001-5437-4073">Qiusheng Wu</a>)
**Repository:** <a href="https://github.com/giswqs/geemap" target ="_blank">https://github.com/giswqs/geemap</a>
**Version:** v0.7.0
