### 2020-05-24 bc13827
Make Google Maps optional #79
