+++ geemap/geemap.py
@@ -70,7 +70,7 @@ class Map(ipyleaflet.Map):
             kwargs['zoom'] = zoom
 
         if 'add_google_map' not in kwargs.keys():
-            kwargs['add_google_map'] = True  
+            kwargs['add_google_map'] = True
 
         # Inherits the ipyleaflet Map class
         super().__init__(**kwargs)
@@ -4577,15 +4577,14 @@ def ee_search(out_repo_dir=None):
 
     tree_widget = widgets.Output()
 
-
     left_widget = widgets.VBox()
     right_widget = widgets.VBox()
     output_widget = widgets.Output()
     output_widget.layout.max_width = '650px'
 
-    repo_tree, repo_tree_dict = build_repo_tree(out_repo_dir, right_widget)
-    left_widget.children = [search_type, search_box, tree_widget]
-    right_widget.children = [output_widget]
+    repo_tree, repo_output, repo_dict = build_repo_tree()
+    left_widget.children = [search_type, repo_tree]
+    right_widget.children = [repo_output]
 
     search_widget = widgets.HBox()
     search_widget.children = [left_widget, right_widget]
@@ -4608,6 +4607,8 @@ def ee_search(out_repo_dir=None):
         tree_widget.clear_output()
         if change['new'] == 'Scripts':
             search_box.placeholder = 'Filter scripts...'
+            left_widget.children = [search_type, repo_tree]
+            right_widget.children = [repo_output]
             with tree_widget:
                 tree_widget.clear_output()
                 print('Loading...')
@@ -4615,6 +4616,8 @@ def ee_search(out_repo_dir=None):
                 display(repo_tree)
         elif change['new'] == 'Docs':
             search_box.placeholder = 'Filter methods...'
+            left_widget.children = [search_type, search_box, tree_widget]
+            right_widget.children = [output_widget]
             with tree_widget:
                 tree_widget.clear_output()
                 print('Loading...')
@@ -4623,6 +4626,8 @@ def ee_search(out_repo_dir=None):
                 right_widget.children = [output_widget]
         elif change['new'] == 'Assets':
             search_box.placeholder = 'Filter assets...'
+            left_widget.children = [search_type, search_box, tree_widget]
+            right_widget.children = [output_widget]
             with tree_widget:
                 tree_widget.clear_output()
                 print('Loading...')
@@ -4707,79 +4712,110 @@ def build_asset_tree():
     return tree, tree_dict
 
 
-def build_repo_tree(out_dir, out_widget):
+def build_repo_tree(out_dir=None, name='gee_repos'):
+    """Builds a repo tree for GEE account.
 
+    Args:
+        out_dir (str): The output directory for the repos. Defaults to None.
+        name (str, optional): The output name for the repo directory. Defaults to 'gee_repos'.
+
+    Returns:
+        tuple: Returns a tuple containing a tree widget, an output widget, and a tree dictionary containing nodes.
+    """
     import warnings
     warnings.filterwarnings('ignore')
 
     if out_dir is None:
-        out_dir = os.path.join(os.path.expanduser('~'), 'gee_repos')
-        if not os.path.exists(out_dir):
-            os.makedirs(out_dir)
+        out_dir = os.path.join(os.path.expanduser('~'))
 
-    tree = Tree(multiple_selection=False)
-    tree_dict = {}
+    repo_dir = os.path.join(out_dir, name)
+    if not os.path.exists(repo_dir):
+        os.makedirs(repo_dir)
 
-    path_widget = widgets.Text()
-    path_widget.layout.width = '450px'
+    URLs = {
+        'Owner': 'https://earthengine.googlesource.com/{}/default'.format(ee_user_id()),
+        'Writer': '',
[…]
