### 2020-06-08 #2305 @whedon
[REVIEW]: geemap: A Python package for interactive mapping with Google Earth Engine
**Submitting author:** @giswqs (<a href="http://orcid.org/0000-0001-5437-4073">Qiusheng Wu</a>)
**Repository:** <a href="https://github.com/giswqs/geemap" target ="_blank">https://github.com/giswqs/geemap</a>
**Version:** v0.7.6
**Archive:** <a href="https://doi.org/10.6084/m9.figshare.12608834.v1" target="_blank">10.6084/m9.figshare.12608834.v1</a>
