### 2020-06-08 #2305 @hugoledoux
👋🏼 @fbiljecki @steflhermitte this is the review thread for the paper. All of our communications will happen here from now on.

It's for both of you the first review here, so it can be a bit overwhelming. Please feel free to ping me (@hugoledoux) if you have any questions/concerns.
