### 2020-06-12 396104c
Implement Assets Tab
