+++ geemap/geemap.py
@@ -2929,17 +2929,17 @@ def ee_export_image_to_drive(ee_object, description, folder=None, region=None, s
     try:
         params = {}
 
-        if folder is not None:      
+        if folder is not None:
             params['driveFolder'] = folder
         if region is not None:
-            params['region'] = region 
-        if scale is None:  
-            scale = ee_object.projection().nominalScale().multiply(10) 
-        params['scale'] = scale        
+            params['region'] = region
+        if scale is None:
+            scale = ee_object.projection().nominalScale().multiply(10)
+        params['scale'] = scale
         if crs is not None:
-            params['crs'] = crs        
+            params['crs'] = crs
         params['maxPixels'] = max_pixels
-        params['fileFormat'] = file_format  
+        params['fileFormat'] = file_format
 
         task = ee.batch.Export.image(ee_object, description, params)
         task.start()
@@ -2962,12 +2962,12 @@ def ee_export_image_collection_to_drive(ee_object, descriptions=None, folder=Non
         crs (str, optional): CRS to use for the exported image.. Defaults to None.
         max_pixels (int, optional): Restrict the number of pixels in the export. Defaults to 1.0E13.
         file_format (str, optional): The string file format to which the image is exported. Currently only 'GeoTIFF' and 'TFRecord' are supported. Defaults to 'GeoTIFF'.
-    """  
+    """
     ee_initialize()
 
     if not isinstance(ee_object, ee.ImageCollection):
         print('The ee_object must be an ee.ImageCollection.')
-        return  
+        return
 
     try:
         count = int(ee_object.size().getInfo())
@@ -2975,8 +2975,8 @@ def ee_export_image_collection_to_drive(ee_object, descriptions=None, folder=Non
 
         if (descriptions is not None) and (len(descriptions) != count):
             print('The number of descriptions is not equal to the number of images.')
-            return     
-        
+            return
+
         if descriptions is None:
             descriptions = ee_object.aggregate_array('system:index').getInfo()
 
@@ -2985,10 +2985,11 @@ def ee_export_image_collection_to_drive(ee_object, descriptions=None, folder=Non
         for i in range(0, count):
             image = ee.Image(images.get(i))
             name = descriptions[i]
-            ee_export_image_to_drive(image, name, folder, region, scale, crs, max_pixels, file_format)
+            ee_export_image_to_drive(
+                image, name, folder, region, scale, crs, max_pixels, file_format)
 
     except Exception as e:
-        print(e)    
+        print(e)
 
 
 def ee_to_numpy(ee_object, bands=None, region=None, properties=None, default_value=None):
@@ -4640,23 +4641,38 @@ def search_api_tree(keywords, api_tree):
     return sub_tree
 
 
-def ee_search(out_repo_dir=None):
+def ee_search(asset_limit=100):
     """Search Earth Engine API and user assets. If you received a warning (IOPub message rate exceeded) in Jupyter notebook, you can relaunch Jupyter notebook using the following command:
         jupyter notebook --NotebookApp.iopub_msg_rate_limit=10000
-    """
+
+    Args:
+        asset_limit (int, optional): The number of assets to display for each asset type, i.e., Image, ImageCollection, and FeatureCollection. Defaults to 100.
+    """    
+
     import warnings
     warnings.filterwarnings('ignore')
+
+    class Flags:
+        def __init__(self, repos=None, docs=None, assets=None, docs_dict=None, asset_dict=None, asset_import=None):
+            self.repos = repos
+            self.docs = docs
+            self.assets = assets
+            self.docs_dict = docs_dict
+            self.asset_dict = asset_dict
+            self.asset_import = asset_import
+
+    flags = Flags()
 
[…]
