### 2020-06-12 cb4fdf2
Add 300+ geemap examples
