### 2020-06-15 70bd297
Update py template
