+++ HISTORY.rst
@@ -1,6 +1,8 @@
 =======
 History
 =======
+0.7.0 (2020-05-22)
+------------------
 0.6.0 (2020-04-05)
 ------------------
 0.5.0 (2020-03-24)
