+++ setup.py
@@ -61,6 +61,6 @@ setup(
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/giswqs/geemap',
-    version='0.7.1',
+    version='0.7.2',
     zip_safe=False,
 )
