### 2020-06-19 786e039
Fix encoding bug #97
