### 2020-06-23 #2305 @hugoledoux
@fbiljecki I see you ticked several checkboxes above, but not all. If you haven't done some yet it's fine, but if you have questions or think the software/docs should be modified, the idea is that you state so here in a new comment. If it's something complex you can open a new issue in the repo of the project and put a link here too.
