### 2020-06-23 #2305 @giswqs
@fbiljecki Thank you very much for taking the time to review the repo! Regarding the two unchecked items:

- **Performance**: There are no performance claims of the package.
- **Automated tests**: I have some automated tests in [tests/test_geemap.py](https://github.com/giswqs/geemap/blob/master/tests/test_geemap.py) as well as [py-check.yaml](https://github.com/giswqs/geemap/blob/master/.github/workflows/py-check.yaml). Would these be sufficient?

I would be happy to make improvement if needed. Thank you.
