+++ geemap/geemap.py
@@ -87,6 +87,8 @@ class Map(ipyleaflet.Map):
         self.draw_layer = None
         self.draw_last_json = None
         self.draw_last_bounds = None
+        self.user_roi = None
+        self.user_rois = None
 
         self.plot_widget = None  # The plot widget for plotting Earth Engine data
         self.plot_control = None  # The plot control for interacting plotting
@@ -345,12 +347,14 @@ class Map(ipyleaflet.Map):
             try:
                 self.draw_count += 1
                 geom = geojson_to_ee(geo_json, False)
+                self.user_roi = geom
                 feature = ee.Feature(geom)
                 self.draw_last_json = geo_json
                 self.draw_last_bounds = minimum_bounding_box(geo_json)
                 self.draw_last_feature = feature
                 self.draw_features.append(feature)
                 collection = ee.FeatureCollection(self.draw_features)
+                self.user_rois = collection
                 ee_draw_layer = ee_tile_layer(
                     collection, {'color': 'blue'}, 'Drawing Features', True, 0.5)
 
@@ -369,6 +373,7 @@ class Map(ipyleaflet.Map):
                 self.draw_features = []
                 self.draw_last_feature = None
                 self.draw_layer = None
+                self.user_roi = None
 
         draw_control.on_draw(handle_draw)
         self.add_control(draw_control)
@@ -2520,6 +2525,15 @@ def open_youtube():
     webbrowser.open_new_tab(url)
 
 
+def api_docs():
+    """Open a browser and navigate to the geemap API documentation.
+    """
+    import webbrowser
+
+    url = 'https://geemap.readthedocs.io/en/latest/source/geemap.html#geemap-package'
+    webbrowser.open_new_tab(url)
+
+
 def show_youtube(id='h0pz3S6Tvx0'):
     """Displays a YouTube video within Jupyter notebooks.
 
@@ -2753,6 +2767,39 @@ def ee_export_vector(ee_object, filename, selectors=None):
         print('Data downloaded to {}'.format(filename))
     except Exception as e:
         print(e)
+
+
+def ee_export_vector_to_drive(ee_object, description, folder, file_format='shp', selectors=None):
+    """Exports Earth Engine FeatureCollection to Google Drive. other formats, including shp, csv, json, kml, and kmz.
+
+    Args:
+        ee_object (object): ee.FeatureCollection to export.
+        description (str): File name of the output file.
+        folder (str): Folder name within Google Drive to save the exported file.
+        file_format (str, optional): The supported file format include shp, csv, json, kml, kmz, and TFRecord. Defaults to 'shp'.
+        selectors (list, optional): The list of attributes to export. Defaults to None.
+    """
+    if not isinstance(ee_object, ee.FeatureCollection):
+        print('The ee_object must be an ee.FeatureCollection.')
+        return
+
+    allowed_formats = ['csv', 'json', 'kml', 'kmz', 'shp', 'tfrecord']
+    if not (file_format.lower() in allowed_formats):
+        print('The file type must be one of the following: {}'.format(
+            ', '.join(allowed_formats)))
+        return
+
+    task_config = {
+        'folder': folder,
+        'fileFormat': file_format,
+    }
+
+    if selectors is not None:
+        task_config['selectors'] = selectors
+
+    print('Exporting {}...'.format(description))
+    task = ee.batch.Export.table.toDrive(ee_object, description, **task_config)
+    task.start()
 
 
 def ee_to_shp(ee_object, filename, selectors=None):
@@ -4086,8 +4133,8 @@ def minimum_bounding_box(geojson):
         bounds = (lower_left, upper_right)
[…]
