### 2020-06-26 3dce17c
Fix Windows encoding bug #97
