### 2020-06-29 172047f
Improve utils module
