+++ geemap/utils.py
@@ -8,62 +8,283 @@ except Exception as e:
     ee.Initialize()
 
 
-# Compute area in square meters
-def vec_area(f):
-    # Compute area in square meters.  Convert to hectares.
-    areaSqm = f.area()
-    # A new property called 'area' will be set on each feature.
-    return f.set({'area': areaSqm})
+def vec_area(fc):
+    """Calculate the area (m2) of each each feature in a feature collection.
+
+    Args:
+        fc (object): The feature collection to compute the area.
+
+    Returns:
+        object: ee.FeatureCollection
+    """
+    return fc.map(lambda f: f.set({'area_m2': f.area(1).round()}))
 
 
-def vec_area_sqkm(f):
-    areaSqkm = f.area().divide(1000 * 1000)
-    return f.set({'area': areaSqkm})
+def vec_area_km2(fc):
+    """Calculate the area (km2) of each each feature in a feature collection.
+
+    Args:
+        fc (object): The feature collection to compute the area.
+
+    Returns:
+        object: ee.FeatureCollection
+    """
+    return fc.map(lambda f: f.set({'area_km2': f.area(1).divide(1e6).round()}))
 
 
-def vec_area_ha(f):
-    # Compute area in square meters.  Convert to hectares.
-    areaHa = f.area(1).divide(100 * 100)
+def vec_area_mi2(fc):
+    """Calculate the area (square mile) of each each feature in a feature collection.
 
-    # A new property called 'area' will be set on each feature.
-    return f.set({'area': areaHa})
+    Args:
+        fc (object): The feature collection to compute the area.
+
+    Returns:
+        object: ee.FeatureCollection
+    """
+    return fc.map(lambda f: f.set({'area_mi2': f.area(1).divide(2.59e6).round()}))
 
 
-def get_year(date):
-    return ee.Date(date).get('year')
+def vec_area_ha(fc):
+    """Calculate the area (hectare) of each each feature in a feature collection.
+
+    Args:
+        fc (object): The feature collection to compute the area.
+
+    Returns:
+        object: ee.FeatureCollection
+    """
+    return fc.map(lambda f: f.set({'area_ha': f.area(1).divide(1e4).round()}))
 
 
-# Convert string to number
-def str_to_number(str):
+def image_cell_size(img):
+    """Retrieves the image cell size (e.g., spatial resolution)
+
+    Args:
+        img (object): ee.Image
+
+    Returns:
+        float: The nominal scale in meters.
+    """
+    return img.projection().nominalScale().getInfo()
+
+
+def image_date(img, date_format='YYYY-MM-dd'):
+    """Retrieves the image acquisition date.
+
+    Args:
+        img (object): ee.Image
+        date_format (str, optional): The date format to use. Defaults to 'YYYY-MM-dd'.
+
+    Returns:
+        str: A string representing the acquisition of the image.
+    """
+    return ee.Date(img.get('system:time_start')).format(date_format).getInfo()
+
+
+def image_area(img, region=None, scale=None, denominator=1.0):
+    """Calculates the the area of an image.
+
+    Args:
+        img (object): ee.Image
+        region (object, optional): The region over which to reduce data. Defaults to the footprint of the image's first band.
+        scale (float, optional): A nominal scale in meters of the projection to work in. Defaults to None.
+        denominator (float, optional): The denominator to use for converting size from square meters to other units. Defaults to 1.0.
+
[…]
