### 2020-06-30 8edd9c7
Add more utils functions
