### 2020-06-30 1d583c2
Add more utils functions
