### 2020-07-02 907a4f5
Improve xarray installation
