### 2020-07-02 #2305 @steflhermitte
@hugoledoux Practical question: how do I tick the review checklist? I see the checkboxes, but can't check them.
