### 2020-07-02 #2305 @giswqs
> Dear @giswqs,
>
> First I would like to thank you for the development of what looks at first sight/experience an extremely useful, powerful and well-designed package for interacting with GEE. Looking forward to explore it further in the coming days when finishing the review.
>
> When installing the package I faced some initial glitches as `pip install geemap` failed to correctly install the ipyleaflet and ipytree dependencies (on macports + pip). They should have been automatically installed (I did receive any error message), but I could not load them in my notebook. I managed to solve it by manually re-installing these packages. I don't think it is necessarily a geemap issue, but it could provide a hurdle for others. I don't know how to handle this though in your package description and/or if it is needed, but wanted to mention it.
>
> Hope to finish the rest of the review in the coming hours/days.
>
> Groet,
> Stef

@steflhermitte Thank you very much for taking to time to test the package. The ipyleaflet dependency sometime can be a bit tricky to install, as it might conflict with some of you existing packages. The best way to install geemap is to create a fresh conda environment, if you have Anaconda or Miniconda installed on your computer. I also have a [YouTube video](https://youtu.be/h0pz3S6Tvx0) showing users how to install the package using conda. Hope it helps. Thanks.

https://github.com/giswqs/geemap#installation

```
conda create -n gee python=3.7
conda activate gee
conda install mamba -c conda-forge
mamba install geemap -c conda-forge
```
