### 2020-07-02 #2305 @giswqs
> @hugoledoux Practical question: how do I tick the review checklist? I see the checkboxes, but can't tick them.

You might have received an invitation email earlier on. You will need to click the link in that email to accept the invitation. After that, you should be able to tick the review checklist.
