### 2020-07-02 #2305 @steflhermitte
@giswqs Thanks for the clarifications! I agree that your manual is very clear. I was just testing it in my own environment and that seems to work (after solving the conflicting dependencies).

The invitation seems not to work any longer `Sorry, we couldn't find that repository invitation It is possible that the invitation was revoked or that you are not logged into the invited account.` I am logged in with the correct account, so I guess that I need to be re-invited to be able to tick the boxes.
