### 2020-07-02 #2305 @steflhermitte
Dear @giswqs and @hugoledoux,

I have been reviewing the software and accompanying paper and want to congratulate @giswqs on the nice piece of software. Geemap contains very useful tools for the designed audience that want to use Google Earth Engine in a Jupyter Notebook environment. The tools are simple, perform good and are well documented by example notebooks and accompanying Youtube videos.

Based on this review I recommend the acceptance of this submission.

Best regards,
Stef Lhermitte
