### 2020-07-03 #2305 @hugoledoux
Nice work @giswqs, this review was very smooth! The 2 reviewers now recommend acceptance so we're reaching the end of the review!

At this point could you:

- [ ] accept the PR I made to your paper, there was a small typo in the acknowlegment
- [ ] Archive the reviewed software in Zenodo or a similar service (e.g. figshare, an institutional repository)
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.

Please list the DOI of the archived version here.
