+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.7.5
+current_version = 0.7.6
 commit = True
 tag = True
 
