### 2020-07-03 #2305 @giswqs
Dear @hugoledoux, I have made the requested changes. Thanks.

1. Merged your pull request.
2. Made a tagged version [v0.7.6](https://github.com/giswqs/geemap/releases/tag/v0.7.6)
3. Archived the software in [figshare](https://doi.org/10.6084/m9.figshare.12608834.v1)
4. Figshare citation: Wu, Q., 2020. geemap: A Python package for interactive mapping with Google Earth Engine. https://doi.org/10.6084/m9.figshare.12608834.v1
