### 2020-07-04 #2305 @FainazInamdeen
i have installed geemap through Anaconda, but when I coding in Jupiter notebook, I cannot open the map as you shown in video, what would be the wrong?
