### 2020-07-04 #2305 @giswqs
@FainazInamdeen Did you create a fresh conda env to install geemap? Check the [installation guidelines](https://github.com/giswqs/geemap#installation). If you still encounter issues, please open a new issue [here](https://github.com/giswqs/geemap/issues). This thread is for the joss-review paper only, not for trouble shooting package installation issues.

```
conda create -n gee python=3.7
conda activate gee
conda install mamba -c conda-forge
mamba install geemap -c conda-forge
```
