### 2020-07-04 #2305 @FainazInamdeen
@Qiusheng Wu
Sorry didn't notice, btw I have sorted the issue. I think it was because
jupyter notebook before I opened through IE after I changed it into google
chrome, now its working. Thank you for an amazing tutorial.

On Sat, Jul 4, 2020 at 10:38 PM Qiusheng Wu <notifications@github.com>
wrote:

> @FainazInamdeen <https://github.com/FainazInamdeen> Did you create a
> fresh conda env to install geemap? Check the installation guidelines
> <https://github.com/giswqs/geemap#installation>. If you still encounter
> issues, please open a new issue here
> <https://github.com/giswqs/geemap/issues>. This thread is for the
> joss-review paper only, not for trouble shooting package installation
> issues.
>
> conda create -n gee python=3.7
> conda activate gee
> conda install mamba -c conda-forge
> mamba install geemap -c conda-forge
>
