### 2020-07-05 61fec1c
Fix duplicate layer bug
