+++ geemap/eefolium.py
@@ -20,6 +20,7 @@ def ee_initialize():
         ee.Authenticate()
         ee.Initialize()
 
+
 # More WMS basemaps can be found at https://viewer.nationalmap.gov/services/
 ee_basemaps = {
     'ROADMAP': folium.TileLayer(
@@ -116,7 +117,7 @@ ee_basemaps = {
         name='Esri National Geographic',
         overlay=True,
         control=True
-    ),     
+    ),
 
     'Esri Shaded Relief': folium.TileLayer(
         tiles='https://services.arcgisonline.com/arcgis/rest/services/World_Shaded_Relief/MapServer/tile/{z}/{y}/{x}',
@@ -124,7 +125,7 @@ ee_basemaps = {
         name='Esri Shaded Relief',
         overlay=True,
         control=True
-    ),   
+    ),
 
     'Esri Physical Map': folium.TileLayer(
         tiles='https://services.arcgisonline.com/arcgis/rest/services/World_Physical_Map/MapServer/tile/{z}/{y}/{x}',
@@ -132,7 +133,7 @@ ee_basemaps = {
         name='Esri Physical Map',
         overlay=True,
         control=True
-    ),   
+    ),
 
     'Bing VirtualEarth': folium.TileLayer(
         tiles='http://ecn.t3.tiles.virtualearth.net/tiles/a{q}.jpeg?g=1',
@@ -144,34 +145,39 @@ ee_basemaps = {
 
     '3DEP Elevation': folium.WmsTileLayer(
         url='https://elevation.nationalmap.gov/arcgis/services/3DEPElevation/ImageServer/WMSServer?',
-        layers = '3DEPElevation:None',
+        layers='3DEPElevation:None',
         attr='USGS',
         name='3DEP Elevation',
         overlay=True,
         control=True
-    ),    
+    ),
 
     'NAIP Imagery': folium.WmsTileLayer(
         url='https://services.nationalmap.gov/arcgis/services/USGSNAIPImagery/ImageServer/WMSServer?',
-        layers = '0',
+        layers='0',
         attr='USGS',
         name='NAIP Imagery',
         overlay=True,
         control=True
-    ),           
+    ),
 }
 
 
 class Map(folium.Map):
     """The Map class inherits from folium.Map. By default, the Map will add Google Maps as the basemap. Set add_google_map = False to use OpenStreetMap as the basemap.
-    
+
     Args:
         folium (object): An folium map instance.
 
     Returns:
         object: folium map object.
-    """    
+    """
+
     def __init__(self, **kwargs):
+
+        import logging
+        logging.getLogger('googleapiclient.discovery_cache').setLevel(logging.ERROR)
+        ee_initialize()
 
         # Default map center location and zoom level
         latlon = [40, -100]
@@ -192,10 +198,10 @@ class Map(folium.Map):
         if 'zoom_start' in kwargs.keys():
             zoom = kwargs['zoom_start']
         else:
-            kwargs['zoom_start'] = zoom 
+            kwargs['zoom_start'] = zoom
 
         if 'add_google_map' not in kwargs.keys():
-            kwargs['add_google_map'] = True              
+            kwargs['add_google_map'] = True
 
         super().__init__(**kwargs)
 
@@ -205,7 +211,6 @@ class Map(folium.Map):
         # plugins.Fullscreen().add_to(self)
 
         self.fit_bounds([latlon, latlon], max_zoom=zoom)
- 
 
     def setOptions(self, mapTypeId='HYBRID', styles={}, types=[]):
         """Adds Google basemap to the map.
@@ -214,30 +219,30 @@ class Map(folium.Map):
[…]
