+++ README.rst
@@ -133,6 +133,12 @@ To use **geemap**, you must first `sign up <https://earthengine.google.com/signu
   conda install mamba -c conda-forge
   mamba install geemap -c conda-forge 
 
+Optionally, you can install `Jupyter notebook extensions <https://github.com/ipython-contrib/jupyter_contrib_nbextensions>`__, which can improve your productivity in the notebook environment. Some useful extensions include Table of Contents, Gist-it, Autopep8, Variable Inspector, etc. See this `post <https://towardsdatascience.com/jupyter-notebook-extensions-517fa69d2231>`__ for more information.       
+
+.. code:: python
+
+  mamba install jupyter_contrib_nbextensions -c conda-forge 
+
 
 If you have installed **geemap** before and want to upgrade to the latest version, you can run the following command in your terminal:
 
@@ -145,7 +151,7 @@ If you use conda, you can update geemap to the latest version by running the fol
   
 .. code:: python
 
-  conda update -c conda-forge geemap
+  mamba update -c conda-forge geemap
 
 
 To install the development version from GitHub using `Git <https://git-scm.com/>`__, run the following command in your terminal:
