+++ docs/geemap.rst
@@ -1,0 +1,70 @@
+geemap package
+==============
+
+Submodules
+----------
+
+geemap.basemaps module
+----------------------
+
+.. automodule:: geemap.basemaps
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+geemap.cli module
+-----------------
+
+.. automodule:: geemap.cli
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+geemap.conversion module
+------------------------
+
+.. automodule:: geemap.conversion
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+geemap.eefolium module
+----------------------
+
+.. automodule:: geemap.eefolium
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+geemap.geemap module
+--------------------
+
+.. automodule:: geemap.geemap
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+geemap.legends module
+---------------------
+
+.. automodule:: geemap.legends
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+geemap.utils module
+-------------------
+
+.. automodule:: geemap.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: geemap
+   :members:
+   :undoc-members:
+   :show-inheritance:
