### 2020-07-05 83fbd0d
Add support for publishing maps #109
