### 2020-07-05 086230e
Add support for publishing maps #109
