+++ geemap/__init__.py
@@ -2,7 +2,7 @@
 
 __author__ = """Qiusheng Wu"""
 __email__ = 'giswqs@gmail.com'
-__version__ = '0.7.6'
+__version__ = '0.7.7'
 
 from .geemap import *
 from .basemaps import ee_basemaps
