+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.7.6
+current_version = 0.7.7
 commit = True
 tag = True
 
