### 2020-07-06 #2305 @hugoledoux
Congratulations @giswqs  your submission is now accepted! Your paper will be handled and published in the near future.

Thanks to @fbiljecki  and @steflhermitte  for the reviews and feedback!
