### 2020-07-06 #2305 @giswqs
@hugoledoux @fbiljecki @steflhermitte Thank you very much for taking the time to review the package. Much appreciated.
