+++ geemap/utils.py
@@ -1,9 +1,18 @@
 
 """Module of various Earth Engine utility functions.
 """
+import os
 import ee
 
 try:
+    ee_token = os.environ.get('EARTHENGINE_TOKEN')
+    if ee_token is not None:
+        credential = '{"refresh_token":"%s"}' % ee_token
+        credential_file_path = os.path.expanduser("~/.config/earthengine/")
+        os.makedirs(credential_file_path, exist_ok=True)
+        with open(credential_file_path + 'credentials', 'w') as file:
+            file.write(credential)
+
     ee.Initialize()
 except Exception as e:
     ee.Authenticate()
