### 2020-07-07 1535a4a
Bump version: 0.7.7 → 0.7.8
