+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.7.7
+current_version = 0.7.8
 commit = True
 tag = True
 
