### 2020-07-07 6ca487c
Add attribution control (#111) and JOSS paper
