+++ geemap/utils.py
@@ -4,19 +4,23 @@
 import os
 import ee
 
-try:
-    ee_token = os.environ.get('EARTHENGINE_TOKEN')
-    if ee_token is not None:
-        credential = '{"refresh_token":"%s"}' % ee_token
-        credential_file_path = os.path.expanduser("~/.config/earthengine/")
-        os.makedirs(credential_file_path, exist_ok=True)
-        with open(credential_file_path + 'credentials', 'w') as file:
-            file.write(credential)
+def ee_initialize(token_name='EARTHENGINE_TOKEN'):
+    """Authenticates Earth Engine and initialize an Earth Engine session
 
-    ee.Initialize()
-except Exception as e:
-    ee.Authenticate()
-    ee.Initialize()
+    """
+    try:
+        ee_token = os.environ.get(token_name)
+        if ee_token is not None:
+            credential = '{"refresh_token":"%s"}' % ee_token
+            credential_file_path = os.path.expanduser("~/.config/earthengine/")
+            os.makedirs(credential_file_path, exist_ok=True)
+            with open(credential_file_path + 'credentials', 'w') as file:
+                file.write(credential)
+
+        ee.Initialize()
+    except:
+        ee.Authenticate()
+        ee.Initialize()
 
 
 def vec_area(fc):
