### 2020-07-07 eaf7ca6
Add attribution control (#111) and JOSS paper
