### 2020-07-10 b294900
Improve landsat timelapse
