### 2020-07-10 1fac9a3
Improve landsat timelapse
