+++ geemap/geemap.py
@@ -81,7 +81,7 @@ class Map(ipyleaflet.Map):
             kwargs['add_google_map'] = True
 
         if 'show_attribution' not in kwargs.keys():
-            kwargs['show_attribution'] = True           
+            kwargs['show_attribution'] = True
 
         # Inherits the ipyleaflet Map class
         super().__init__(**kwargs)
@@ -321,13 +321,14 @@ class Map(ipyleaflet.Map):
 
         self.add_control(control=data_control)
 
-        search_marker = Marker(icon=AwesomeIcon(name="check", marker_color='green', icon_color='darkgreen'))
-        search = SearchControl(position="topleft", 
-                       url='https://nominatim.openstreetmap.org/search?format=json&q={s}', 
-                       zoom=5,
-                       property_name='display_name',
-                       marker=search_marker
-                      )
+        search_marker = Marker(icon=AwesomeIcon(
+            name="check", marker_color='green', icon_color='darkgreen'))
+        search = SearchControl(position="topleft",
+                               url='https://nominatim.openstreetmap.org/search?format=json&q={s}',
+                               zoom=5,
+                               property_name='display_name',
+                               marker=search_marker
+                               )
         self.add_control(search)
 
         self.add_control(ZoomControl(position='topleft'))
@@ -1731,7 +1732,7 @@ class Map(ipyleaflet.Map):
             print(e)
             return
 
-    def add_landsat_ts_gif(self, layer_name='Timelapse', roi=None, label=None, start_year=1984, end_year=2019, start_date='06-10', end_date='09-20', bands=['NIR', 'Red', 'Green'], vis_params=None, dimensions=768, frames_per_second=10, font_size=30, font_color='black', add_progress_bar=True, progress_bar_color='white', progress_bar_height=5, out_gif=None):
+    def add_landsat_ts_gif(self, layer_name='Timelapse', roi=None, label=None, start_year=1984, end_year=2019, start_date='06-10', end_date='09-20', bands=['NIR', 'Red', 'Green'], vis_params=None, dimensions=768, frames_per_second=10, font_size=30, font_color='black', add_progress_bar=True, progress_bar_color='white', progress_bar_height=5, out_gif=None, upload=False):
         """Adds a Landsat timelapse to the map.
 
         Args:
@@ -1751,7 +1752,8 @@ class Map(ipyleaflet.Map):
             add_progress_bar (bool, optional): Whether to add a progress bar at the bottom of the GIF. Defaults to True.
             progress_bar_color (str, optional): Color for the progress bar. Defaults to 'white'.
             progress_bar_height (int, optional): Height of the progress bar. Defaults to 5.
-            out_gif ([type], optional): File path to the output animated GIF. Defaults to None.
+            out_gif (str, optional): File path to the output animated GIF. Defaults to None.
+            upload (bool, optional): Whether to upload the gif to imgur.com. Defaults to False.
 
         """
         try:
@@ -1788,14 +1790,19 @@ class Map(ipyleaflet.Map):
                 add_text_to_gif(in_gif, in_gif, xy=('2%', '90%'), text_sequence=label,
                                 font_size=font_size, font_color=font_color, duration=int(1000 / frames_per_second), add_progress_bar=add_progress_bar, progress_bar_color=progress_bar_color, progress_bar_height=progress_bar_height)
 
+            reduce_gif_size(in_gif)
+
             bounds = minimum_bounding_box(geojson)
             # bounds = ((35.892718, -115.471773), (36.409454, -114.271283))
[…]
