### 2020-07-10 e877ae9
Bump version: 0.7.8 → 0.7.9
