+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.7.8
+current_version = 0.7.9
 commit = True
 tag = True
 
