### 2020-07-15 #2305 @hugoledoux
@giswqs no you don't, it should be picked up by one of the journal editor. It's summer I guess, sorry for delay
