### 2020-07-15 1d4a381
minor changes in paper
