+++ paper/paper.md
@@ -30,21 +30,21 @@ phenology [@Li2019], and time series analysis [@Kennedy2018].
 GEE provides both JavaScript and Python APIs for making computational requests to the Earth Engine servers.
 Compared with the comprehensive [documentation](https://earthengine.google.com/) and interactive IDE
 (i.e., [GEE JavaScript Code Editor](https://code.earthengine.google.com/)) of the GEE JavaScript API, the
-GEE Python API lacks good documentation and functionality for visualizing results interactively.
+GEE Python API lacks good documentation and lacks functionality for visualizing results interactively.
 The **geemap** Python package is created to fill this gap. It is built upon
 [ipyleaflet](https://github.com/jupyter-widgets/ipyleaflet) and [ipywidgets](https://github.com/jupyter-widgets/ipywidgets),
 enabling GEE users to analyze and visualize Earth Engine datasets interactively with Jupyter notebooks.
 
 # geemap Audience
 
-**geemap** is intended for students and researchers, who would like to utilize the Python ecosystem of diverse libraries and
+**geemap** is intended for students and researchers who would like to utilize the Python ecosystem of diverse libraries and
 tools to explore Google Earth Engine. It is also designed for existing GEE users who would like to transition from the GEE
-JavaScript API to Python API. The automated JavaScript-to-Python [conversion module](https://github.com/giswqs/geemap/blob/master/geemap/conversion.py)
+JavaScript API to a Python API. The automated JavaScript-to-Python [conversion module](https://github.com/giswqs/geemap/blob/master/geemap/conversion.py)
 of the **geemap** package can greatly reduce the time needed to convert existing GEE JavaScripts to Python scripts and Jupyter notebooks.
 
 # geemap Functionality
 
-The interactive mapping functionality of the **geemap** package is built upon [ipyleaflet](https://github.com/jupyter-widgets/ipyleaflet) and [folium](https://github.com/python-visualization/folium), both of which rely on Jupyter notebooks for creating interactive maps. A key difference between ipyleaflet and folium is that ipyleaflet is built upon ipywidgets and allows bidirectional communication between the front-end and the backend enabling the use of the map to capture user input, while folium is meant for displaying static data only [@QuantStack2019]. It should be noted that [Google Colab](https://colab.research.google.com/) currently does not support ipyleaflet. Therefore, if one wants to use **geemap** on Google Colab, one should use `import geemap.eefolium as geemap`, which provides limited interactive mapping functionality. To utilize the full interactive mapping functionality of **geemap**, one should use `import geemap` on a local computer or secured server with Jupyter notebook installed.
+The interactive mapping functionality of the **geemap** package is built upon [ipyleaflet](https://github.com/jupyter-widgets/ipyleaflet) and [folium](https://github.com/python-visualization/folium), both of which rely on Jupyter notebooks for creating interactive maps. A key difference between ipyleaflet and folium is that ipyleaflet is built upon ipywidgets and allows bidirectional communication between the frontend and the backend, enabling the use of the map to capture user input, while folium is meant for displaying static data only [@QuantStack2019]. It should be noted that [Google Colab](https://colab.research.google.com/) currently does not support ipyleaflet. Therefore, if one wants to use **geemap** on Google Colab, one should `import geemap.eefolium as geemap`, which provides limited interactive mapping functionality. To utilize the full interactive mapping functionality of **geemap**, one should `import geemap` on a local computer or secured server with Jupyter notebook installed.
 
 The key functionality of **geemap** is organized into several modules:
 
[…]
