### 2020-07-15 a36c121
minor fixes in bib, adding DOI
