+++ paper/paper.bib
@@ -3,11 +3,11 @@
                everyone}",
   author    = "Gorelick, Noel and Hancher, Matt and Dixon, Mike and
                Ilyushchenko, Simon and Thau, David and Moore, Rebecca",
-  journal   = "Remote sensing of environment",
+  journal   = "Remote Sensing of Environment",
   publisher = "Elsevier",
   volume    =  202,
   pages     = "18--27",
-  month     =  "1~" # dec,
+  month     =  dec,
   year      =  2017,
   url       = "http://www.sciencedirect.com/science/article/pii/S0034425717302900",
   keywords  = "Cloud computing; Big data; Analysis; Platform; Data
@@ -18,18 +18,17 @@
 
 @ARTICLE{Hansen2013,
   title     = "{High-resolution global maps of 21st-century forest cover change}",
-  author    = "Hansen, M C and Potapov, P V and Moore, R and Hancher, M and
-               Turubanova, S A and Tyukavina, A and Thau, D and Stehman, S V
-               and Goetz, S J and Loveland, T R and Kommareddy, A and Egorov, A
-               and Chini, L and Justice, C O and Townshend, J R G",
+  author    = "Hansen, M. C. and Potapov, P. V. and Moore, R. and Hancher, M. and
+               Turubanova, S. A. and Tyukavina, A. and Thau, D. and Stehman, S. V.
+               and Goetz, S. J. and Loveland, T. R. and Kommareddy, A. and Egorov, A.
+               and Chini, L. and Justice, C. O. and Townshend, J. R. G.",
   journal   = "Science",
   publisher = "science.sciencemag.org",
   volume    =  342,
   number    =  6160,
   pages     = "850--853",
-  month     =  "15~" # nov,
+  month     =  nov,
   year      =  2013,
-  url       = "http://dx.doi.org/10.1126/science.1244693",
   language  = "en",
   issn      = "0036-8075, 1095-9203",
   pmid      = "24233722",
@@ -40,13 +39,13 @@
   title     = "{Google Earth Engine, {Open-Access} Satellite Data, and Machine
                Learning in Support of {Large-Area} Probabilistic Wetland
                Mapping}",
-  author    = "Hird, Jennifer N and DeLancey, Evan R and McDermid, Gregory J
+  author    = "Hird, Jennifer N. and DeLancey, Evan R. and McDermid, Gregory J.
                and Kariyeva, Jahan",
   publisher = "Multidisciplinary Digital Publishing Institute",
   volume    =  9,
   number    =  12,
   pages     = "1315",
-  month     =  "14~" # dec,
+  month     =  dec,
   year      =  2017,
   url       = "http://www.mdpi.com/2072-4292/9/12/1315",
   keywords  = "VIP;Earth Engine",
@@ -57,15 +56,15 @@
 @ARTICLE{Kennedy2018,
   title     = "{Implementation of the LandTrendr Algorithm on Google Earth
                Engine}",
-  author    = "Kennedy, Robert E and Yang, Zhiqiang and Gorelick, Noel and
-               Braaten, Justin and Cavalcante, Lucas and Cohen, Warren B and
+  author    = "Kennedy, Robert E. and Yang, Zhiqiang and Gorelick, Noel and
+               Braaten, Justin and Cavalcante, Lucas and Cohen, Warren B. and
                Healey, Sean",
   journal   = "Remote Sensing",
   publisher = "Multidisciplinary Digital Publishing Institute",
   volume    =  10,
   number    =  5,
   pages     = "691",
-  month     =  "1~" # may,
+  month     =  may,
   year      =  2018,
[…]
