### 2020-07-15 #2305 @danielskatz
@giswqs - Please merge the changes in https://github.com/giswqs/geemap/pull/115 then we can proceed to publication
