### 2020-07-15 #2305 @danielskatz
Thanks to @fbiljecki & @steflhermitte for reviewing and @hugoledoux for editing!

Congratulations to @giswqs (Qiusheng Wu)!!
